Project Changes
Script Changes:
(N) BaseStat.cs: Provides the base class to be inherited by other classes.
(N) PlayerStat.cs: Inherits from BaseStat. It handles all the stat for the player.
(M) BaseMovement.cs: Removed Start() and Update() in the script as it is a abstract class.
(M) PlayerMovementCC.cs: Added regions for the script for organization.